Tested EasyGetTextPairsStrategy.cs
Made a test and fixed the GetRandomPairs() function. Issue with previous implementation is when use JsonFs without closing after getting len_letters, because we don't save the RootElement object but instead try to parse json file again, unable to recognize json file as JsonFs is somewhere in middle of file. So added a line to move back to beginning of stream.

Signed-off-by: FDIL <fdil.fadilh1@gmail.com>