Abstracted grid creation function.
Because each game grid creation is essentially the same, except for number of sides, decided to extract that to a function that takes the number of sides as a parameter. Then have specific data templates for each game board just call a function that calls the abstracted function with number of sides.

Signed-off-by: FDIL <fdil.fadilh1@gmail.com>